+++ test/src/version_test.dart
@@ -1,0 +1,13 @@
+import 'dart:io';
+
+import 'package:dapper/src/version.dart';
+import 'package:test/test.dart';
+import 'package:yaml/yaml.dart';
+
+void main() {
+  test('packageVersion matches pubspec.yaml', () {
+    final pubspec =
+        loadYaml(File('pubspec.yaml').readAsStringSync()) as YamlMap;
+    expect(packageVersion, pubspec['version']);
+  });
+}